Protected all the routes

+++ Api/app.js
@@ -8,6 +8,9 @@ const bodyParser = require('body-parser');
 //Load Mongoose Models
 const { List, Task, User } = require('./db/models');
 
+
+const jwt = require('jsonwebtoken');
+
 /* MIDDLEWARE */
 
 //Load Middleware
@@ -26,12 +29,31 @@ app.use(function (req, res, next) {
     next();
 });
 
+//check whethe the request has a Valid JWT access token
+let authenticate = (req, res, next) => {
+    let token = req.header('x-access-token');
+
+    //verify the JWT
+    jwt.verify(token, User.getJWTSecret(), (err, decoded) => {
+        if (err) {
+            //If jwt is INVALID do not Authenticate
+            res.status(401).send(err);
+        } else {
+            //Jwt is Valid
+            req.user_id = decoded._id;
+            next();
+        }
+    })
+}
+
+
+//Verify Refresh Token MiddleWare (which will be verifying the session)
 let verifySession = (req, res, next) => {
     let refreshToken = req.header('x-refresh-token');
     let _id = req.header('_id');
 
     User.findByIdAndToken(_id, refreshToken).then((user) => {
-        if(!user) {
+        if (!user) {
             return Promise.reject({
                 'error': 'User not Found. Make sure that the refresh Token and User id are correct'
             });
@@ -46,18 +68,18 @@ let verifySession = (req, res, next) => {
         let isSessionValid = false;
 
         user.sessions.forEach((session) => {
-            if(session.token === refreshToken ) {
-                if(User.hasRefreshTokenExpired(session.expiresAt) === false ){
+            if (session.token === refreshToken) {
+                if (User.hasRefreshTokenExpired(session.expiresAt) === false) {
                     isSessionValid = true;
                 }
             }
         });
 
-        if(isSessionValid) {
+        if (isSessionValid) {
             //Session is valid - call next() to continue processing this web request 
             next();
         }
-        else{
+        else {
             return Promise.reject({
                 'error': 'Refresh Token has expired or the session is invalid'
             })
@@ -76,20 +98,25 @@ let verifySession = (req, res, next) => {
  * Purpose: Get all lists from DB
  */
 
-app.get('/lists', (req, res) => {
-    // Return an array of all the list in the DB
-    List.find({}).then((lists) => {
+app.get('/lists', authenticate, (req, res) => {
+    // Return an array of all the list in the DB that belong to the authenticated user
+    List.find({
+        _userId: req.user_id
+    }).then((lists) => {
         res.send(lists);
+    }).catch((e) => {
+        res.send(e);
     });
 })
 
-app.post('/lists', (req, res) => {
+app.post('/lists', authenticate, (req, res) => {
     // Create a new List and retur the same to the use with its ID
     // The list info will be passed on by the JSON request body
     let title = req.body.title;
 
     let newList = new List({
-        title
+        title,
+        _userId: req.user_id
     });
     newList.save().then((listDoc) => {
         //Full list document is returned
@@ -97,28 +124,31 @@ app.post('/lists', (req, res) => {
     })
 });
 
-app.patch('/lists/:id', (req, res) => {
+app.patch('/lists/:id', authenticate, (req, res) => {
     // We will update the specified list with the new values specified in th JSON body
-    List.findOneAndUpdate({ _id: req.params.id }, {
+    List.findOneAndUpdate({ _id: req.params.id, _userId: req.user_id }, {
         $set: req.body
     }).then(() => {
         res.sendStatus(200);
     });
 });
 
-app.delete('/lists/:id', (req, res) => {
+app.delete('/lists/:id', authenticate, (req, res) => {
     // We want to delete the specified list
     List.findOneAndRemove({
-        _id: req.params.id
+        _id: req.params.id,
+        _userId: req.user_id
     }).then((removedListDoc) => {
         res.send(removedListDoc);
+        //delte all the task that are in the list
+        deleteTaskFromList(removedListDoc._id);
     })
 })
 
 /**
  * Get all tasks in a Specific list
  */
-app.get('/lists/:listId/tasks', (req, res) => {
+app.get('/lists/:listId/tasks', authenticate, (req, res) => {
     Task.find({
         _listId: req.params.listId
     }).then((tasks) => {
@@ -129,28 +159,62 @@ app.get('/lists/:listId/tasks', (req, res) => {
 /**
  * Create a new task in a specified List
  */
-app.post('/lists/:listId/tasks', (req, res) => {
-    let newTask = new Task({
-        title: req.body.title,
-        _listId: req.params.listId
-    });
-    newTask.save().then((newTaskDoc) => {
-        res.send(newTaskDoc);
+app.post('/lists/:listId/tasks', authenticate, (req, res) => {
+    List.findOne({
+        _id: req.params.listId,
+        _userId: req.user_id
+    }).then((list) => {
+        if (list) {
+            //Valid Object so currently authenticated user can create tasks
+            return true;
+        }
+        else {
+            return false;
+        }
+    }).then((canCreateTask) => {
+        if (canCreateTask) {
+            let newTask = new Task({
+                title: req.body.title,
+                _listId: req.params.listId
+            });
+            newTask.save().then((newTaskDoc) => {
+                res.send(newTaskDoc);
+            })
+        }
+        else {
+            res.sendStatus(404);
+        }
     })
+
 })
 
 
 /**
  * Update an existing task
  */
-app.patch('/lists/:listId/tasks/:taskId', (req, res) => {
-    Task.findOneAndUpdate({
-        _id: req.params.taskId,
-        _listId: req.params.listId
-    }, {
-        $set: req.body
-    }).then(() => {
-        res.send({ message: 'Updated Successfully' });
+app.patch('/lists/:listId/tasks/:taskId', authenticate, (req, res) => {
+
+    List.findOne({
+        _id: req.params.listId,
+        _userId: req.user_id
+    }).then((list) => {
+        if (list) {
+            return true;
+        }
+        return false;
+    }).then((canUpdateTasks) => {
+        if (canUpdateTasks) {
+            Task.findOneAndUpdate({
+                _id: req.params.taskId,
+                _listId: req.params.listId
+            }, {
+                $set: req.body
+            }).then(() => {
+                res.send({ message: 'Updated Successfully' });
+            })
+        } else {
+            res.sendStatus(404);
+        }
     })
 });
 
@@ -158,13 +222,28 @@ app.patch('/lists/:listId/tasks/:taskId', (req, res) => {
 /**
  * Delete a task
  */
-app.delete('/lists/:listId/tasks/:taskId', (req, res) => {
-    Task.findOneAndRemove({
-        _id: req.params.taskId,
-        _listId: req.params.listId
-    }).then((removedTaskDoc) => {
-        res.send(removedTaskDoc);
-    })
+app.delete('/lists/:listId/tasks/:taskId', authenticate, (req, res) => {
+    List.findOne({
+        _id: req.params.listId,
+        _userId: req.user_id
+    }).then((list) => {
+        if (list) {
+            return true;
+        }
+        return false;
+    }).then((canDeleteTasks) => {
+        if (canDeleteTasks) {
+            Task.findOneAndRemove({
+                _id: req.params.taskId,
+                _listId: req.params.listId
+            }).then((removedTaskDoc) => {
+                res.send(removedTaskDoc);
+            })
+        }
+        else {
+            res.sendStatus(404);
+        }
+    });
 });
 
 /* User Routes */
@@ -243,6 +322,14 @@ app.get('/users/me/access-token', verifySession, (req, res) => {
 })
 
 
+/* Helper Method */
+let deleteTaskFromList = (_listId) => {
+    Task.deleteMany({
+        _listId
+    }).then(() => {
+        console.log("Tasks from " + _listId + "were Deleted");
+    })
+}
 
 
 app.listen(3000, () => {
